modelagem dre; fks em imovel; app dados_comuns
