+++ sme_ofertaimoveis/dados_comuns/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from .models import DiretoriaRegional, Secretaria
+
+admin.site.register(DiretoriaRegional)
+admin.site.register(Secretaria)
