+++ sme_ofertaimoveis/dados_comuns/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DadosComunsConfig(AppConfig):
+    name = "sme_ofertaimoveis.dados_comuns"
