+++ sme_ofertaimoveis/users/models.py
@@ -3,7 +3,8 @@ import uuid as uuid
 from django.contrib.auth.models import AbstractUser
 from django.db import models
 from django.urls import reverse
-from rest_framework.exceptions import ValidationError
+
+from sme_ofertaimoveis.dados_comuns.models import DiretoriaRegional, Secretaria
 
 
 class Perfil(models.Model):
@@ -17,21 +18,11 @@ class Perfil(models.Model):
         verbose_name_plural = "Perfis"
 
 
-class Secretaria(models.Model):
-    nome = models.CharField("Nome", max_length=30, null=True, blank=True)
-
-    def __str__(self):
-        return f"{self.nome}"
-
-    class Meta:
-        verbose_name = "Secretaria"
-        verbose_name_plural = "Secretarias"
-
-
 class User(AbstractUser):
     uuid = models.UUIDField(default=uuid.uuid4, editable=False, unique=True)
-    perfil = models.ForeignKey(Perfil, null=True, on_delete=models.SET_NULL)
-    secretaria = models.ForeignKey(Secretaria, null=True, on_delete=models.SET_NULL)
+    perfil = models.ForeignKey(Perfil, null=True, blank=True, on_delete=models.SET_NULL)
+    secretaria = models.ForeignKey(Secretaria, null=True, blank=True,  on_delete=models.SET_NULL)
+    dre = models.ForeignKey(DiretoriaRegional, null=True, blank=True, on_delete=models.SET_NULL)
 
     def get_absolute_url(self):
         return reverse("users:detail", kwargs={"username": self.username})
